Import of FLORA/Catalyst-View-Mason-0.14 from CPAN.
gitpan-cpan-distribution: Catalyst-View-Mason
gitpan-cpan-version:      0.14
gitpan-cpan-path:         FLORA/Catalyst-View-Mason-0.14.tar.gz
gitpan-cpan-author:       FLORA
gitpan-cpan-maturity:     released

+++ t/helper.t
@@ -45,6 +45,8 @@ ok(!TestApp::View::Mason->config->{use_match}, 'module sets use_match to false')
 ok(TestApp::View::Mason->isa('Catalyst::View::Mason'), 'module inherits from C::V::Mason');
 
 END {
-    chdir $old_cwd;
-    rmtree($test_dir);
+    if (defined $old_cwd) {
+        chdir $old_cwd;
+        rmtree($test_dir);
+    }
 }
